Yesterday I installed Cuda 9.0, but I could not figure out how to pass the right arguments to the compiler.
Most likely, what I will have to do is install the C++ components of VS so I can run the Cuda example and have a look at the template it is using to call the Cuda compiler.

I have no idea why NVCC is not working. For some reason it cannot find cl even though I am passing the path to it directly.

Well, let me get on with it instead of arguing on Reddit as interesting as the argument I am having with a Lisper is.