https://stackoverflow.com/questions/47375540/what-is-the-correct-template-to-call-nvcc-cuda-9-0-from-the-command-line
I've been busy with this, but I finally figured out what all the problems are and now it works fine.

One thing that bothers me currently is that the parser is rather slow. Discounting the startup time for th epevaller, I think the parser is now slower than it.

Let me spend some time optimizing it then.