While I wait for lunch to finish, let me do some thinking out loud.
Today the way I wasted time is quite remarkable. Instead of fiddling with the parser I should have profiled it first. In the end I optimized the number parser and that turned out to have done completely nothing.

And inlining those parsers actually slowed things down most likely because strings stated manifesting. I tried inlining the pattern parsers, but that did absolutely nothing to improve performance.

The idea to move the parsing and prepass for the libraries is a very good one - there is just one problem, I have absolutely no way of doing that. I got ahead of myself here.

Actually, this realization is pretty significant. I never realized it before, but this is what in essence separates Racket from .NET. If I want to play with staging and complete the language and give it great libraries, Racket does have a better ecosystem for that sort of thing.

Moving to Racket would in essence give me not what I wrote above, but also access to multi-staged computation.

I'd need to write the Cuda wrappers myself, but I am really only using such a tiny piece of the libraries at any given time.

The same is true for .NET so having to look into all those assemblies just to do interop is such a pain.

.NET is really not doing it for me.

Now that I've come this far, this is where my eye on the future is telling me things are going.

But not now. I want to finish the language in this form and I want to finish the ML library. The rest can come after I establish my base. Yeah, I know I say that and then wander of for months to do the thing I say I would not do now, but rewriting the language in Racket is a tall stretch right now.

Then really what does come next here?

I have the map and I have the allocator. I guess I do need the general matrix multiply. Before I call it a day for now, let me make a pass through the old Spiral library to see what I am missing.

How about I forget backwards pass now and focus only on making it go forward all the way to the end. I also need to bring in the cuRand as well.