When I try to run it again, I get that stupid method not found error. I thought that was a Repl thing, but it seems it is also a .NET assembly thing.
Ugh, forget this. This is giving me such a world of pain.

Yeah, although I have got the interop to work, it is a constant upwards battle with .NET. Just like with C++, I am going to go to an all macro based approach towards interop.

I simply made a mistake of ever doing it this way.

The good news is that taking all that stuff out of the language will simplify it markedly.