You know what I will do? I think I will just make the damn types take the same input as the macros do. That way the codegen will know right away what to print. Brilliant isn't it?
I had enough for the day. Tomorrow I will have a lot of examples to redo, but it won't be too many. I am not really using the .NET types in too great of a fashion. I am using the StringBuilder in places and bunch of other stuff. It won't take me longer than tomorrow to get through it along with CudaRandom.

What I am missing in the language for it currently are enums.

What a pain, what a pain. First it was the entension methods, then the method error in the REPL and now this. Also I get the same error in compiled when I repeated run `learning`. It says that the method can't be found.

I can't keep up with .NET's bullshit.

Without a doubt, I am going to beat this and complete the language. But while I do that I will study Racket on the side.

When it comes time to draw out the full power of AIs, I am going to need the capacity for multi-stage compilation without a doubt.