I got the test15 to work now. I'll do the rest later today.
I am not sure whether this is the right choice, but it is one I will be taking. Ultimately what I want out of Spiral is a DL library, not to compete with F#.

Also being able to do this sort of interop is unique to Spiral so I should take advantage of it.

Without a doubt, I will complete the library. Forget the parser, forget the compiler speed, forget the lack of ergonomics. I can't afford to take any compromises on either language speed or expressiveness. Eventually, I will figure out how to take Spiral to the next level, the most important thing is that I solidify my base.