Done with the tests. The conversion to Extern macros turned to be easier than I'd thought. In one fell swoop, I solved the interop problem for two wildly disparate languages. This is how first class staging + intensional polymorphism should be used.
Doing this kind of thing is something that is exclusively the province of Spiral.

Besides this, I've also let the naked_type be printed as errors since it can be difficult to know where they appear otherwise.

Today's changes mark a change in policy of Spiral. No longer will I make the language shoulder the burden of typechecking on its own. F# and NVCC can shoulder some of that weight as well.

Honestly, the effort of trying to make the language polished is too much for me. I am not even going to try from here on out.

Now what was the plan? To put in random numbers. I do not feel like programming much right now, but let me at least catch up to where I was before.