Well, here is random. For all their verbosity, the macros deal with the interop stuff cleanly in the end.
In the long run, I'll be better off not depending on inbuilt reflection.

Now the partial evaluation part of this takes only 0.2 down from 0.6 earlier.

Damn why are compilers so slow these days? It is ridiculous. I did not even realize that NVCC needed 0.9s to finish.

Today this conversion took me forever and it was not particularly easy. Programming in Spiral is really difficult, especially refactoring.

Fuuu, but I got the random numbers down sure enough.

Tomorrow I will use this to initialize the weights and then I will write the 1d sum and put in gemm after that. It might be doable to do it all tomorrow.

The backwards pass will have to wait a bit longer, but it won't be long now until I get logistic regression working on Mnist.

One by one, I will revive the old library, much stronger than before.

But for now, I think I will call it a day here. Got to unwind.