Ok, I finally put in the CudaRandom functions. This is quite nice. I can make do with this.
Unlike in the previous version of the library, I can take advantage of intensional polymorphism to make it generic for the 4 different distributions.

Last time that did not even occur to me. I really went through great pains to make anything of it before.

Ok, now that I have this much let me take a breather. I definitely need it after this marathon session. After that I will do the 1d reduce.

After that, most likely tomorrow I will move into putting in the gemm. That will be enough to do the forward pass for the neural net.

After that I will seriously spend some time on figuring out how to customize the IDE so that when I press F5 it builds and then run that thing first.

I'll ask the guy on the F# sub for advice.