set_unsafe out' i (f (index_unsafe in' i))
I am going insane here. I just realized that in' and out' were reversed before and now that I've corrected them of course the program stops working. The most miraculous thing is that this worked at all.

I am going to have to dive into the guts of this.

...

"3:55pm. Ah, I get it.

Let me try something. If my hunch is right I am probably passing the arguments in the reverse order.

4pm. Yeah, I am passing them in the reverse order. Great. Let me fix this.

4:05pm. What a bug."