I decided to do some refactoring and once again I've run into a ultra hard type error.
No forget it. The language is a failure at this rate.

Rather than finishing the library, I'd be better off shotting myself, that is how difficult programming in Spiral currently is. it is insane.

Just what the hell does this pattern miss mean?

I have no fucking idea.

...

Ah, wait, wait. It has to be this thing.
| {size=({from to} & dim_range) :: size ar}, i :: is ->

I am assuming it will be from to, but it always could be something like a standard int.