I got the map_redo example to work.
In various ways today was absolutely tremendous. Doing proper error reporting instead of AST dumps will be a significant boost to my producstivity.

So far I've been programming solely in shitty Spiral and have seen the ugly side of it. The tedious type error and the great effort it took to write anything will now vanish.

With this I can do it. I can draw out the full power of the language. I look forward to seeing how the workflow will change from here.

Ahhh, damn I am tired. It is way too late to be working on this. Tomorrow I will focus on putting in gemm.