Right now I am studying Racket stuff and I do not feel like programming yet.
I am trying to determine what the future of the language should be after I am done with the library so let me take some time today to deal with that.

...No it is not Racket. What I need to do is implement it in a dynamic language, but I also need that language to make it easy to turn objects into types, and serialize them, and determine whether they are immutable.

I am slowly forming an idea of what is needed, but let me study for now until I get an itch out of the way.