No, I can't really get into Racket. I am simply not into it. I recognize the value of macros and compile time computation, but I can be bothered to switch sides so completely. It is just too much for me.
I am looking at some of the other articles on Tomassetti's site and they are pretty boring.

http://www.doc.ic.ac.uk/~svb/TSfPL/notes.pdf
Type Systems for Programming Languages

https://gitlab.inria.fr/fpottier/mpri-2.4-public/blob/master/README.md
Functional programming and type systems

Let me study this for a while. Type systems still interest me.

I just do not feel like programming today. Rather I am trying to refine Spiral's future.

Honestly, both Racket's and MetaOcaml's approaches to staging fail to catch my interest.

I've also been thinking of how to take advantage of .NET's dynamic features. Going in that direction would be a lot more promising, but I would need some help to handle it. DLR is a ghost town and doing proper .NET interop really takes diving into its innards.

Probably I am going to delay the next version of Spiral with full first class staging until something comes along to break the stalemate and offer me an acceptable choice of platform.

Webassembly? That could fit the bill.

It is not like what I want to do is that urgent. I mean having to do the parsing and the prepass for the libraries will increase the compile times, but not the runtime peformance.

The platform is really the bottleneck for Spiral right now. Also I could really make good use of Lancet, but that does not exist for the .NET either, only JVM.

...

Ok, I guess if I want to stick with ML that only really leaves Ocaml as a viable target.

I think I will sit on this for a year and once the language is a bit more established ask around for a bit.

That will also allow the Ocaml guys to do better Windows support in the meantime. Or WASM will turn out to be something viable. Or maybe MS will come up with something to help language devs.

Only good things can happen by being lazy here. It is a bit similar to the Futhark situation.

Regarding Futhark, even if gets a Cuda backend I won't judge it by the same lax criteria as I did before. I am absolutely going to require it to be a value add if it want to be a partner rather than a preferential experiment like I had in mind previously.

Enough of this.

Let me study the type system stuff for the rest of the day and I will gather my will for the rest of the move. I will not think about the futrue alternatives after today - my only goals will be to finish and then crush.

Unless something good comes along, Spiral is good enough for me in its current state.