I am finally ready to start.
No more hesitation, no more looking back. No remorse, no regret. Deus vult.

https://gitlab.inria.fr/fpottier/mpri-2.4-public/blob/master/README.md

Those lecture notes from were interesting. I had no idea ML& existed which allowed pattern matching on types at runtime, but those have overhead.

The section on compiler optimizations was interesting as well. With this I can firmly state that in some segments Spiral's capabilities are firmly beyond state of the art. Expecially with that list last example. I was definitely struck with the feeling that those guys have no idea what they are doing as I read that.

I've been thinking about future extensions to the language for the past two days. I've been thinking of the semantics when Lit is any arbitrary object.

My conclussion is that unless that object is immutable, there is no way to let it cross the phase boundary. At first I thought it being copyable would be enough, but that is not so.

It would be possible to use mutable objects during compile time, but in order to type them and have them be abstractly interpreted they absolutely need to be immutable.

There is no other choice in that regard.

Really, this is giving me great trouble.

Spiral is a good first step towards the Singularity, and I can think of all sorts of ways to extend the language and make it more flexible and powerful, but those things, even very simple ones feel just so out of my reach.

So I won't let it bother me anymore.

I tried briefly to get into Racket again, but for a metaprogramming based language it is not entirely satisfactory.

A proper metaprogramming language would be something between Haskell and F#...actually, it would be pretty much like Spiral, except not the kind of Spiral I am making right now which uses macro hacks for interop, but one which has a platform and a JIT specifically tailored for it and has access to dynamic typing and is able to differentiate between pure and impure objects and serialize them. That would be necessary for them to cross the phase boundary.

Anything else would be a pain.

In other words, Spiral needs more capabilities to talk to its clones. I simply do not have the prerequisites to move past this and none of the platforms are really suitable for it.

Most likely, in its current incarnation Spiral will remain the pinnacle of pre-Singularity languages.

It will simply have to make do. If I cannot do it with this, then nothing will be enough.

If I need some specific compile time functions like for loading files into strings and generating random numbers, I think I will just add them to the language directly as Ops.

It is painful, but that will be the plan.

Now let me do the forward pass. I will do this and after I train a feedforward net on MNIST for the first time, I will take a break from working on the language to make tutorials for it.