Yeah, this is it. This is the Spiral's true power.
Errors are actually good now and trivial to fix when I miss an argument.

Let me make an adjustmnet to error reporting.

I keep getting the error in the wrong place when I accidentally return a non-unit, so let me put the error there as the doctor orders.

For some reason the application was not the place I was looking for.