I am done for the day. I got the Mnist to load. Tomorrow, what I will do is assert the tensor sizes so they can be statically determined. That is a thing only available to Spiral. Besides that, I need to chunk the arrays into minibatches before loading them onto the GPU.
That in itself will have an interesting challenge regarding how to fit that last unaligned minibatch.

Yeah, this is a bit more fun the usual fare. These kinds of programming days ideal.

Tomorrow I will spend some time on setting up the Mnist module, and if I get lucky, I will start work on the backwards pass.

I am full of determination. Maybe by the time the New Year comes around and the time comes for the next update, I will have something to show for it.

Maybe? No, this time it is certain.