"10:20am. While I was in bed, I did a lot of thinking what I wanted to do today.
Futhark has a really high level of abstraction, but for myself I need to aim differently. Tensors are not nested arrays and I can't really figure out how to fit them into such framework. What I really need instead would be init.

A large majority of things can be expressed in terms of init.

I'll forget about the Cuda side for a bit and go back to host tensor. The first thing I will do is curry the arguments to it. That will be very useful as it will allow me to program in terms of loops instead of having a single one. This will no doubt break all existing code.

Also there is another idea what just came to me right now.

Suppose I have a tensor of sizes (a,b,c). Instead of requiring set and index to specify all the dimensions all at once, I would be better off emulating partial application via views.

Yeah, that would be great. On the Cuda side that would not be hard since I can just push pointers around, but on the .NET side, I'd need to keep an offset around. Well, I could probably make use of such a scheme on Cuda side as well. But I'd rather leave it implicit, so I guess I'll just wrap the array then instead.

I can see it.

Also I think I will change {from to} in sizes to {from near_to}. This later way is better since I won't have to add 1 and substract it to calculate dimension sizes."

I said that today I would work on loading Mnist data, but as a prerequite to that, in order to make dealing with tensors easy, I need to work more on my tensor abstractions.

I think partially applying init and tensor setters and getters would be a spectacular feature.

Doing something like this never even occurred in the previous version of the library in F#. It is no wonder, since what I want to do here would all be impossible in F#.

This idea of using tensor views by partial application is absolutely brilliant and would be significant gain in simplicity and flexibility of my code. It is worth dropping everything to get it done.

I'll take it easily as usual.

Today I got up relatively late, so I will have to start that way as well. Also I am interested how Julia deals with tensors so I'll take a look at that as well.

Let me have breakfast for now.