Ok, apart from show, I think this should be enough for the new HostTensor. The design of it right now is so brilliant that it is blinding me.
This thing I wrote just now is easy some of my best stuff. A vast improvement over the previous HostTensor design.

I am going to replace it now and then spend some time going through all the tests again. This will take a while as I am going to have to resubmit some of the HackerRank stuff just to be thorough.

I'll do it in a new branch.