I removed test85 and that hard rewrite. If I have need for it, I'll just do it by hand. It is not worth having it if it necessitates putting hacks elsewhere in order to have it around.
I do not think this is quite it for destructure, but it should be good enough for the time being.

I am glad that doing this gave me a chance to find and fix a compile bug and simplify the whole thing by removing that never used if_dynamic statement.

Right now, the normal if does not even exist in the language anymore. Good ridance.

Now that I've made the language meaner and leaner once more, I think I am about ready to move to the next steps.

I had a lot of time to think of how I am going to do the Mnist data munging.

Having the new tensor in will make it quite easy to implement things such as views and other operations so instead of chunking it in different arrays, I will be able to keep it all together in one place. Good.

I have two very important pieces now, which is the new tensor and the things it allows me to do and that will be the focus of my effort next.

The other part that I am still thinking about are those abstractions I am still missing for Cuda kernels. I've been thinking about them for a while now and have no good answer, but let me put the aside for now.

Today was full of distraction for me and I really could not get as much done as I'd have liked.

Let me slowly get into the grove for the next part. I need to put in views and reshaping directly into the tensor.

Also I should not assume that offset has to exist. On the Cuda side it might be better to just push the pointer around, but I will stick to the offset for now. No need to make things more complicated than they are already.

Actually, what I will do then doing the to_dev transformation is set the offset to literal 0 and the pointer to the new value. That way I could afford to not send the thing through the function call.

Brilliant. This way is ideal.

Really, getting those unused arguments down to literals is the ideal since it would minimize register pressure. Once Cuda spills, it dumps them into global memory.

...For now instead of just deliberating on this commit, how about I put in views and get the Mnist loading out of the way. I will start off by asserting the expected dimensions then.