I am ready to start for the day.
I did a bunch of figuring out during the night. I know the semantics of CTFE now and the difference between it and the runtime JITing. I might need to do CTFE at some point, but doing it at runtime is pretty useless as the compiler has no access to type information. Well, it might be possible to do it via a annotation as in macros or by returning a dynamic type, but I am not interested in it. If I had to do it, it would be via annotation.

CTFE would be easy as all I need to do is merge the partial evaluation and the codegen phases and find a way to link with the F# compiler.

https://github.com/epfldata/squid/

I've been looking into Scala's Squid and this is probably the final thing that made me decide that no other frameworks for metaprogramming apart from Spiral are worth learning. I would put Lisp/Scheme/Racket's systems in the same backend.

I've seen up to a dozen of them up to now and all of them feel like really shitty DSLs plopped onto a base language.

They are not worth even thinking about. I mean Squid allows the programmer to rename free variables? The authors sure have a lot of balls putting that thing in.

Scala as a language is well on its way to becoming the next C++. Quite a dubious distiction.

For today, what I will do is debug what I wrote yesterday and then overload show for tuples and primitive types.

I've been thinking since that start than rather than typeclasses, Spiral is much better suited towards the OO approach.

Since it is so expressive, it can emulate the Smalltalk approach except without any overhead.

Well, I will think of the details as I move along.

Let me get the Mnist loading out of the way.