Today I stayed in bed for a long time it seems. I've been thinking about the tutorials for Spiral.
Today I wanted to get started with learning, but before I burden myself with that, I want to test all the pieces of the new HostTensor module individually. Spiral is not as good as for refactoring as F# is so I want to do it step by step.

Let me do that for a bit.