I just did not want to put this down.
I have absolutely no idea how I did it, but going through the search space of possible designs paid off for me now.

This is it, this is the true design of the Tensor module. Resembles ML a little doesn't it now?

OO was false after all. I could not get it to handle immutable updates gracefully. It would have required more and more boilerplate as it scaled. A pity.

In HostTensorPrimitives I wanted to introduce more hierarchy, but somehow that only brought confusion.

In the end, very loose coupling is what makes the most sense and will allow me to switch out the module primitives segment in the Cuda part.

Let me really call it a day here. I do hope I will finish the HostTensor tomorrow.