I will call it a day here. I just did not feel like it after the last commit and decided to just lay in bed.
Getting debugging done will not be an easy task. Getting it all into shape tomorrow will take me a few hours and I have no energy for that now.

Neither am I super motivated to go through debugging in general. I wish Spiral had global type inference as it'd take care of it in a jiffy.

Well, at least it is not as bad as before. It is definitely bearable now.

I'll use the next 15h to build up motivation for the next stretch. For the day, I am out.