I am almost done debugging the thing, but now I have the urge to finally start what I've been meaning to do for a while - show.
I will overload primitive types so that show converts them into strings. That will also allow me to partially evaluate them.

Right now the fact that I cannot display tensors is really a huge pain in the ass and I need to do something about that.

Today is a bit embarassing for me since I got up at 10am and rather than program, I've felt like just reading and watching a video for a few hours.

https://www.youtube.com/watch?v=T6ohwZNL0RQ&list=PLnqUlCo055hW7kU-SBQEhC_87etA5Gqlq&index=16
Assuring AI

This is interesting for me since it kind of samples where the PL crowd is going to go with AI.

It is an interesting direction to be sure, to merge first and second wave methods.

Let me see if I can find some focus today. Let me just get this out of the way.

Since this is taking me a while, I think I will just reply to the NVidia guys to try an empty .cu file.