Ok, that is it for string_concat.
I guess I will call it a day here with this. I still have to adapt the string_concat in Extern, but I do not want to break it before I have the chance to test it.

...Honestly, this seems a lot easeir than it actually is. I definitely do need to decide how I am going to do all of this. Not having Haskell's typeclasses to overload types does make things harder here, but the issue is hardly insurmountable.

I am still thinking about the design.

I think I am just going to let functions be "<unknown>" and deal with it that way rather than throw type errors on encountering them.

...

...

Ok, I think I've decided. I am going to pass in configuration options as lambdas as usual. In arrays I need to do that in order to short circuit on too large inputs.

I am going to have to do something similar for union types.

Printing tensors will not be hard though I will have to implement views just for that purpose.

Thankfully, I will be able to reuse the `show` to print arrays inside it. I won't bother tiling the arrays, but just like for printing code, I will indent them.

Everything will come together nicely.

Let me stop here for the day.

Clearly, this is kind of an annoyance since I want to work on the library rather than on printing functions, but the basic print is so useful and essential that I can't possibly avoid working on this. It won't take too long and then I will be able to start on what I really want to do.

One other thing. Maybe it would be good for me to move Tuple into the CoreLib? I'll think about it.

Also for the sake of tests going faster, I am going to find some way of precompiling the modules at some point.