I am finally done with the thesis. I liked it for the first 100 pages, but in the fexprs were not what I thought they would be.
Based on second hand descriptions of them, I had imagined they would be something like Spiral's staged functions, but the similarities were imaginary. I've long been hoping to find an equivalent language to Spiral done by somebody in order to ground my terminology and borrow ideas, but Kernel's fexprs are not it.

Some of the history stuff in Shutt's thesis is interesting, but nothing there is useful to me apart from expanding my vocabularity and general level of education. This is usual how my studies go.

I really should be focusing on programming more, but programming without study gets tedious really quickly.

Well, let me go through the debugging and then reworking the tuple matching, so I can move on to completing the forward and the backward pass. After that I will be able to start work on the tutorials.