All the tests are finally clear. Phew.And done for the day. I had enough.
It is great that most bugs seem to be of the type variety. They are more difficult to get out than in F#, but once the program compiles it tends to work.

It takes a while to compile now, but it is nothing unusual for its size. I'll see whether it speeds up tomorrow once I improve pattern matching. That is one place where I should focus on speeding things up.

I am thinking, but instead of having that pattern compiler, maybe I should go back to the earlier version of leaving the pattern in and evaling it right there.

This might be something to leave for the future.

Well, back then, before the last blog update, I could not do it, but...

No, the reason why I did it like this in the first place is to significantly simplify patterns.

...I'll think about the idea overnight. I will consider it tomorrow when I am fresher.