I am going to try simpifying the pattern matching on modules radically. I will also get rid of extension patterns. They were a nice experiments, but they aren't what I desire in their current form.
Since the pattern matching for modules is in such a worse state than the lists, I am going to have to do a lot more work. I am not sure whether I will even have time to finish it today, but I will dedicate the next few days to doing it.

Modules are such a major part of Spiral that I am pretty much obligated to do this. Hopefully it should speed up compilation.

Well, let me get on with this.