Try #2. This is how I will significantly improve the design.
I will add more CPS'd operatives, just like TakeN and TakeNTail. It is amazing that this kind of idea never occured to me earier, but there you have it.

Then I will redo the rest of the patterns so use CPS'd operations instead of more expensive ones that create intermediaries.