What I did today was brilliant. The module pattern is done, but I have not tested it yet. Regardless, the amount of simplification I did here is phenomenal.
I am about 75% done with the redesign and what is left won't be too hard, most of the hard stuff I've done today.

I really am a fool for not realizing that I should just CPS the key pattern matching operatives earlier.

I guess it is true that a language is something that never really leaves the alpha phase. I should stop saying I will ever be done.

Well, it is quite late and I worked my ass off, so I will call it a day here. I will be done with this tomorrow and after that, it will be time to make a serious dent in the whole ML library thing.

Now that I got the tensors, there is no reason why development on it should not resume.