I've decided - I won't add any more CPS'd primitives. The module and tuple ones are the exception, not the rule.
The tuple CPS'd operatives speed up the compiler by 25%, and the module operatives simplified it markedly and revealed some bugs.

On the other hand, I could speed up type literals and the rest, but batching the operation like that would increase the complexity of the compiler. I am not up for it.

Let me merge with main.