Yeah, I had enough for the day. I found quite a few bugs in the last 40m. It is just a matter of exercising all the pathways.
And at any rate, I am done with the forward phase on Mnist. This is not all - I am still missing accuracy, but I will leave that aside for the time being.

I am thinking that for tomorrow, it is finally but surely time for me to start work on the backwards pass. It has been too long hasn't it?

It has been 13.5 months since I threw in the towel on doing layer normalization. Bit by bit, I will catch up to where I left of and this time for sure - I will exceed my limit.