Let me start with this. I woke up 15m ago and have gathered the determination to start instead of spending time on Reddit, but I find myself just thinking instead.
Fusing those maps and reduces will be a kid's game in Spiral.

It all composes very nicely in my mind, so forget that for the time being, it is not an issue..

I might need to do some reorganization...

Ok, here is my goal. Let me make ad_map.

I'll forget the Mnist test and just sketch it out for now.