"5:45pm. Honestly, I feel like calling it a day here. Now that I've gotten into grove after a significant effort, I'll be able to finish this tomorrow.
From now on, instead of keeping an explicit tape, I'll just run these like closures.

I think I will make a monad for keeping track of the nodes on the forward pass.

...The current lack of deadness checks also has me worried a little. I will also have to add something for that. That adjoints on the backwards pass are not the only thing I need to think about.

6pm. Is there a way to make it simpler than the last time?

Yeah, this stuff is complicated, no doubt about it. If only I could make the metalanguage of Spiral impure somehow, that would really help. I could then move these checks to compile time.

I can't believe that even Spiral is getting pressed by the program, but it is true. And those deadness checks did save my ass a few times so I do need them.

Actually, I know how to make it work. I'll give every node a unique name and on the backwards pass, check whether they are in a map. For the temporary adjoints, it would be better to allocate them on the backward pass rather than the forward one.

But do let me do this one at a time. One problem with this kind of scheme is that it will strain the language and will not allow me to make very dynamic graphs. It'd be better if I did not do it for the sake of that.

6:05pm. There is just one thing – instead of plugging the deadness checks right in, I need to aim to do them on a separate layer otherwise it will swell the complexity of the library.

The same goes for all the stuff I am doing in map_back right now.

No way I am doing this in map_redo and everywhere else. I am going to find a way to do all this preprocessing in separate phases.

6:10pm. I said the ML library will be a piece of cake after all the exercise I got on doing the compiler, but that is not the case. Doing all of its parts properly is definitely straining my mind. Today proves that it will be a fitting challenge for me."

I am covering so much new ground here that my past experience is not as helpful as it would have been.

I should have taken it slower today, but even though I did not code much, I definitely did not slack. I've been pushing my mind to try to surmount the block every step of the way.

This is the real deal. This is the heavy difficult programming that Spiral was made for.

I cannot go through this quickly. It will take both much inspiration and much effort, but if I focus and work on it every day, the ML library will fall into place just like everything else. Spiral has the power to give me what I desire.

I have many regrets in life, but for once I will be able to go into the beyond.