To start the day off, I cleaned up the pattern_compile just a bit and added module_filter to the core library.
As usual, I won't force myself. I'll work when I feel like it and rest when I feel like it. I've learned my lesson for good.

Programming is an activity where you are blown around by the winds of inspiration. There is no fighting it.

It is a long cycle of gather and release.

So for now, let me get breakfast and then I will redo the two filter functions so they support modules. And so that they work correctly in addition to that.