"3:45pm. This is giving me a lot to think about.     inl sigmoid hidden_size input_size ret =         inb weight = sigmoid_initializer (input_size, hidden_size)         ret (inl input -> gemm input weight >>= sigmoid)
Should I do it like this? But how will I init the next layer?

Should I do it like this then?

3:55pm.
    inl sigmoid hidden_size next_layer input_size ret =
        inb weight = sigmoid_initializer (input_size, hidden_size)
        inb next_layer = next_layer hidden_size
        ret (inl input -> gemm input weight >>= sigmoid >>= next_layer)

I am just staring at this and thinking. But I like this. I really like this a lot.

This form has such immense potential. Of course the >>= operator here can't be the standard CPS one. I need to define it in AutoDiffOps.

4pm. As an aside, I am really happy I added the inb to the language. It makes everything 10x easier. I could not live without it.

4:05pm. I am just in quiet satisfaction here. Let me commit this and I will start work on >>=. I want to think about this a little more, but I guess there is no point in delaying this. The above is most likely the ideal form and part from using the thing from the previous version of Spiral to do the init, there is not much I can do here.

If I leave out the next_layer argument I won't be able to do the init without using…

Ah, it would be possible, I'd have to modify the function to be able to query the hidden size. Then I could make the combinators for the layers.

...But there is not much point to that.

I'll just modify it to accept weight updates and that would be it."