I am satisfied with this. This is good enough for the layers. Recurent layers will be in a different module. I won't try to mix them needlessly.
Considering how simple the resulting code is, today might be seen as disappointing, but this sort of design is not easy.

I think it took me 2/3rd of the way into 2016 to realize how layers might be composed functionally in the previous library, and up to that point I've been doing functional programming full time. Despite that I still could not figure it out until the inspiration hit me.

Today I had the good foresight to stuble upon the winning design on the first try, but then I threw it away when I realize that cost function would not compose with them.

In the end I realized I threw the towel in too quickly decided to come back to them, but not after unsuccessfuly trying various other alternatives.

In the end, what I have here is what composes the best.

I am close and getting closer. Tomorrow I should be able to finish the backward pass.

Always, the work seems immense, but in the end somehow I always manage to partition it into 5 line functions.

That is skill.