I got sidetracked by this.
When I made my pass I realized that primal, adjoint and make_dual_host are incorrect and had to redesign them.

...I am thinking what I am going to do next.

I think it is time I make the run function for the feedforward layers. It is finally time for the real deal.

I will do this and then put all the pieces together and get debugging.