This is so fucked up.
I can't believe this is so difficult. I am trynig to figure out how to divide by the batch size and am turning up a complete blank on how to do it in a coherent manner.

I won't be able to finish the backwards pass today by the looks of it. I need to think this through.

But I vaguely recall this kind of thing being a pain in the ass last time as well.

...I need to think through.