Here is the run function.
It is always hard going for me at the start, but once I get going, I really get into the groove. Without this, I'd never manage to get this far.

The run function is pretty similar to what I had before, but quite nicer. I like it. Spiral's for function is really a treasure right now.

Right now what I am missing are broadcast operations for the biases and the column and row reduces for that an the accuracy calculations.

I definitely need those before I can consider the backwards pass complete. I also need to make the init kernel and maybe rederive the map ones from that.

But that won't be difficult. The actual Cuda parts are such a pitifully small part of the library compared to what they were before.

What I am going to do today, is an essential piece that I did not mention above. The Sgd optimizer. It is just a simple map (+), no problem.

Should I bring is saxpy or do it as straightforward map module?

Well, let me just do it and I will take a break. I am getting tired from this.